+++ lib/hosamAddition/HttpReqstats/Base.dart
@@ -15,19 +15,7 @@ class ApiBase<ResponseObj> extends StatefulWidget {
   final bool useSkeleton;
   final Widget? skeleton;
 
-  ApiBase({
-    Key? key,
-    required this.requestFunction,
-    HDMHttpRequestsStates<ResponseObj>? httpRequestsStates,
-    required this.buildIdle,
-    required this.buildLoading,
-    required this.buildSuccess,
-    required this.buildError,
-    required this.buildEmptySuccess,
-    this.useSkeleton = false,
-    this.skeleton,
-  })  : httpRequestsStates = httpRequestsStates ?? HDMHttpRequestsStates<ResponseObj>(),
-        super(key: key);
+  ApiBase({Key? key, required this.requestFunction, HDMHttpRequestsStates<ResponseObj>? httpRequestsStates, required this.buildIdle, required this.buildLoading, required this.buildSuccess, required this.buildError, required this.buildEmptySuccess, this.useSkeleton = false, this.skeleton}) : httpRequestsStates = httpRequestsStates ?? HDMHttpRequestsStates<ResponseObj>(), super(key: key);
 
   @override
   State<ApiBase> createState() => _ApiBaseState<ResponseObj>();
@@ -70,10 +58,7 @@ class _ApiBaseState<ResponseObj> extends State<ApiBase<ResponseObj>> {
       builder: (context, snapshot) {
         if (snapshot.connectionState == ConnectionState.waiting) {
           if (widget.useSkeleton && widget.skeleton != null) {
-            return Skeletonizer(
-              enabled: true,
-              child: widget.skeleton!,
-            );
+            return Skeletonizer(enabled: true, child: widget.skeleton!);
           }
           return widget.buildLoading(context);
         } else if (snapshot.hasError) {